requirements.txt
PyQt5
pyautogui
pyperclip
python-socketio
numpy
pygame
Pillow
opencv-python

